backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
supabase==2.0.0
pydantic==2.5.0
pydantic[email]==2.5.0
python-multipart==0.0.6
reportlab==4.0.7
python-dateutil==2.8.2
stripe==7.0.0
email-validator>=2.1.1
requests==2.31.0
Pillow==10.1.0
python-jose[cryptography]==3.3.0
boto3==1.34.0
slowapi==0.1.9

